move variant diff to binf module
